Extract interactive demos into Client Components
Following the RSC pattern, pages are now Server Components that import
compiled Client Components for interactive elements:

- AccordionDemo: AccordionSingleOpenDemo, AccordionMultipleOpenDemo
- CheckboxDemo: CheckboxBindingDemo, CheckboxWithLabelDemo
- CounterDemo: CounterInteractiveDemo, CounterDerivedDemo
- InputDemo: InputBindingDemo, InputFocusDemo
- SwitchDemo: SwitchInteractiveDemo, SwitchSettingsPanelDemo
- TabsDemo: TabsBasicDemo, TabsMultipleDemo, TabsDisabledDemo

Pages now only use "use client" components for interactive parts,
while static examples use base components directly.

Updated E2E tests to use new demo component scope names.
