+++ ui/e2e/switch.spec.ts
@@ -33,17 +33,17 @@ test.describe('Switch Documentation Page', () => {
 
   test.describe('Interactive Toggle', () => {
     test('displays interactive switch', async ({ page }) => {
-      const interactiveSwitch = page.locator('[data-bf-scope="InteractiveSwitch"]').first()
+      const interactiveSwitch = page.locator('[data-bf-scope="SwitchInteractiveDemo"]').first()
       await expect(interactiveSwitch).toBeVisible()
     })
 
     test('shows initial Off state', async ({ page }) => {
-      const interactiveSwitch = page.locator('[data-bf-scope="InteractiveSwitch"]').first()
+      const interactiveSwitch = page.locator('[data-bf-scope="SwitchInteractiveDemo"]').first()
       await expect(interactiveSwitch.locator('text=Off')).toBeVisible()
     })
 
     test('toggles to On state on click', async ({ page }) => {
-      const interactiveSwitch = page.locator('[data-bf-scope="InteractiveSwitch"]').first()
+      const interactiveSwitch = page.locator('[data-bf-scope="SwitchInteractiveDemo"]').first()
       const switchButton = interactiveSwitch.locator('button[role="switch"]')
 
       await switchButton.click()
@@ -60,11 +60,11 @@ test.describe('Switch Documentation Page', () => {
 
   test.describe('Settings Panel', () => {
     test('displays settings panel example', async ({ page }) => {
-      await expect(page.locator('[data-bf-scope="SettingsPanel"]').first()).toBeVisible()
+      await expect(page.locator('[data-bf-scope="SwitchSettingsPanelDemo"]').first()).toBeVisible()
     })
 
     test('shows Wi-Fi, Bluetooth, and Notifications toggles', async ({ page }) => {
-      const settingsPanel = page.locator('[data-bf-scope="SettingsPanel"]').first()
+      const settingsPanel = page.locator('[data-bf-scope="SwitchSettingsPanelDemo"]').first()
       await expect(settingsPanel.locator('text=Wi-Fi')).toBeVisible()
       await expect(settingsPanel.locator('text=Bluetooth')).toBeVisible()
       await expect(settingsPanel.locator('text=Notifications')).toBeVisible()
