+++ ui/pages/tabs.tsx
@@ -1,15 +1,8 @@
-"use client"
 /**
  * Tabs Documentation Page
  */
 
-import { createSignal, createMemo } from '@barefootjs/dom'
-import {
-  Tabs,
-  TabsList,
-  TabsTrigger,
-  TabsContent,
-} from '@/components/Tabs'
+import { TabsBasicDemo, TabsMultipleDemo, TabsDisabledDemo } from '@/components/TabsDemo'
 import {
   PageHeader,
   Section,
@@ -184,168 +177,6 @@ const tabsContentProps: PropDefinition[] = [
   },
 ]
 
-// Interactive example - Basic tabs
-// Note: Uses createMemo to compute selection states because the compiler
-// doesn't yet support inline comparison expressions in JSX props.
-function BasicTabs() {
-  const [activeTab, setActiveTab] = createSignal('account')
-
-  const isAccountSelected = createMemo(() => activeTab() === 'account')
-  const isPasswordSelected = createMemo(() => activeTab() === 'password')
-
-  return (
-    <Tabs value={activeTab()} onValueChange={setActiveTab}>
-      <TabsList>
-        <TabsTrigger
-          value="account"
-          selected={isAccountSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('account')}
-        >
-          Account
-        </TabsTrigger>
-        <TabsTrigger
-          value="password"
-          selected={isPasswordSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('password')}
-        >
-          Password
-        </TabsTrigger>
-      </TabsList>
-      <TabsContent value="account" selected={isAccountSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <h4 class="font-medium mb-2">Account Settings</h4>
-          <p class="text-zinc-500 text-sm">Make changes to your account here. Click save when you're done.</p>
-        </div>
-      </TabsContent>
-      <TabsContent value="password" selected={isPasswordSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <h4 class="font-medium mb-2">Password Settings</h4>
-          <p class="text-zinc-500 text-sm">Change your password here. After saving, you'll be logged out.</p>
-        </div>
-      </TabsContent>
-    </Tabs>
-  )
-}
-
-// Interactive example - Multiple tabs
-function MultipleTabs() {
-  const [activeTab, setActiveTab] = createSignal('overview')
-
-  const isOverviewSelected = createMemo(() => activeTab() === 'overview')
-  const isAnalyticsSelected = createMemo(() => activeTab() === 'analytics')
-  const isReportsSelected = createMemo(() => activeTab() === 'reports')
-  const isNotificationsSelected = createMemo(() => activeTab() === 'notifications')
-
-  return (
-    <Tabs value={activeTab()} onValueChange={setActiveTab}>
-      <TabsList>
-        <TabsTrigger
-          value="overview"
-          selected={isOverviewSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('overview')}
-        >
-          Overview
-        </TabsTrigger>
-        <TabsTrigger
-          value="analytics"
-          selected={isAnalyticsSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('analytics')}
-        >
-          Analytics
-        </TabsTrigger>
-        <TabsTrigger
-          value="reports"
-          selected={isReportsSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('reports')}
-        >
-          Reports
-        </TabsTrigger>
-        <TabsTrigger
-          value="notifications"
-          selected={isNotificationsSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('notifications')}
-        >
-          Notifications
-        </TabsTrigger>
-      </TabsList>
-      <TabsContent value="overview" selected={isOverviewSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">Overview content goes here.</p>
-        </div>
-      </TabsContent>
-      <TabsContent value="analytics" selected={isAnalyticsSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">Analytics content goes here.</p>
-        </div>
-      </TabsContent>
-      <TabsContent value="reports" selected={isReportsSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">Reports content goes here.</p>
-        </div>
-      </TabsContent>
-      <TabsContent value="notifications" selected={isNotificationsSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">Notifications content goes here.</p>
-        </div>
-      </TabsContent>
-    </Tabs>
-  )
-}
-
-// Disabled tab example
-function DisabledTabsExample() {
-  const [activeTab, setActiveTab] = createSignal('active')
-
-  const isActiveSelected = createMemo(() => activeTab() === 'active')
-  const isAnotherSelected = createMemo(() => activeTab() === 'another')
-
-  return (
-    <Tabs value={activeTab()}>
-      <TabsList>
-        <TabsTrigger
-          value="active"
-          selected={isActiveSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('active')}
-        >
-          Active
-        </TabsTrigger>
-        <TabsTrigger
-          value="disabled"
-          selected={false}
-          disabled={true}
-        >
-          Disabled
-        </TabsTrigger>
-        <TabsTrigger
-          value="another"
-          selected={isAnotherSelected()}
-          disabled={false}
-          onClick={() => setActiveTab('another')}
-        >
-          Another
-        </TabsTrigger>
-      </TabsList>
-      <TabsContent value="active" selected={isActiveSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">This tab is active.</p>
-        </div>
-      </TabsContent>
-      <TabsContent value="another" selected={isAnotherSelected()}>
-        <div class="p-4 rounded-lg border border-zinc-200 bg-white">
-          <p class="text-zinc-600">Another active tab.</p>
-        </div>
-      </TabsContent>
-    </Tabs>
-  )
-}
-
 export function TabsPage() {
   return (
     <div class="space-y-12">
@@ -357,7 +188,7 @@ export function TabsPage() {
       {/* Preview */}
       <Example title="" code={`<Tabs>...</Tabs>`}>
         <div class="w-full max-w-md">
-          <BasicTabs />
+          <TabsBasicDemo />
         </div>
       </Example>
 
@@ -376,19 +207,19 @@ export function TabsPage() {
         <div class="space-y-8">
           <Example title="Basic" code={basicCode}>
             <div class="w-full max-w-md">
-              <BasicTabs />
+              <TabsBasicDemo />
             </div>
           </Example>
 
           <Example title="Multiple Tabs" code={multipleTabsCode}>
             <div class="w-full max-w-lg">
-              <MultipleTabs />
+              <TabsMultipleDemo />
             </div>
           </Example>
 
           <Example title="Disabled Tab" code={disabledCode}>
             <div class="w-full max-w-md">
-              <DisabledTabsExample />
+              <TabsDisabledDemo />
             </div>
           </Example>
         </div>
